+++ api/src/pdc/services/policy/engine/policies/20220420_MetropoleRouenNormandie.spec.ts
@@ -2,7 +2,7 @@ import test from 'ava';
 import { v4 } from 'uuid';
 import { OperatorsEnum } from '../../interfaces';
 import { makeProcessHelper } from '../tests/macro';
-import { Mrn as Handler } from './Mrn';
+import { Mrn as Handler } from './20220420_MetropoleRouenNormandie';
 
 const defaultPosition = {
   arr: '76540',
